requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
charset-normalizer==2.0.12
frozenlist==1.3.0
gino==1.0.1
greenlet==1.1.2
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.3.24
typing-extensions==4.1.1
yarl==1.7.2
